Moved Stats service to own namespace
Tidy the root namespace.
