+++ tests/Console/TubeStatsCommandTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Phlib\Beanstalk\Console;
 
 use Phlib\Beanstalk\Exception\InvalidArgumentException;
-use Phlib\Beanstalk\StatsService;
+use Phlib\Beanstalk\Stats\Service as StatsService;
 use PHPUnit\Framework\MockObject\MockObject;
 
 class TubeStatsCommandTest extends ConsoleTestCase
